added algorithm info to docs directory
Though this is code (and can be run) it's more documentation about how
the [simple] algorithm is implemented, hence it being in the docs
directory.

It might make sense to write this up as a markdown file at some point,
and the layout (all the doc strings etc) lends itself to this (as this
was an original intention). For now you can just run / read this
though (it should be quite readable).