Ensure that search doesn't screen bleed or crash
Currently the case where the search is right up against the edge of the
screen isn't handled. It bleeds over into the other side (overlaps like
snake or whatever) and eventually crashes. Rather than this it should
test whether the search is going to go out of range and if so it should
be handled.

Possibly by setting the node to be the edge value - or by just not using
the out of bounds values when displaying nodes. It might more sense to
handle the errors within the graph class than the main edge getting
functions.