searched_around is being used as main - needs work
Using the new methods the main is cleaner now. Though it's not working
properly in that when it gets seeker nodes they seem to have a kind of
'off by 1' effect going on visually (although the algorithm works
alright i think). This is going to be in one of the 'nodes search
around' methods, not sure why it's doing it though.